The most significant changes involve the creation of two new files, `deploy.json` and `deploy.parameters.json`, and the update to the `CRUD application 2.csproj` file to include these new files.
The `deploy.json` file is an Azure Resource Manager (ARM) template for deploying a SQL Server and a SQL Database on Azure. It includes parameters for the server name, database name, admin username, and admin password, and specifies the properties of the SQL Server and the SQL Database.

The `deploy.parameters.json` file is used to provide values for the parameters defined in the `deploy.json` ARM template. The user needs to provide the actual values for these parameters.

Changes:
1. The `CRUD application 2.csproj` file has been updated to include the `deploy.json` and `deploy.parameters.json` files.
2. A new file `deploy.json` has been created. This is an ARM template for deploying a SQL Server and a SQL Database on Azure.
3. A new file `deploy.parameters.json` has been created. This file is used to provide values for the parameters defined in the `deploy.json` ARM template.